+++ xuruko.scss
@@ -1,3 +1,13 @@
+*, *::before, *::after {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+html, body {
+    scroll-behavior: smooth;
+}
+
 :root {
     --red: #f63460;
     --smasherscapeYellow: #fcfc05;
@@ -88,7 +98,7 @@ li {
     padding: 15px 15px 50px;
     grid-template-columns: 1fr;
     justify-content: space-between;
-    @media (min-width: 1366px) {
+    @media (min-width: 1200px) {
         &.populated {
             align-items: flex-start;
             grid-template-columns: 1fr 1fr;
@@ -231,7 +241,7 @@ li {
                             position: relative;
                             bottom: -15px;
                             right: 65px;
-                            @media (min-width: 1366px) and (max-width: 1545px) {
+                            @media (min-width: 1200px) and (max-width: 1545px) {
                                 right: 45px;
                             }
                             @media (max-width: 768px) {
@@ -453,7 +463,7 @@ li {
         @media (min-width: 1500px) and (max-width: 1545px) {
             font-size: 1em;
         }
-        @media (min-width: 1366px) and (max-width: 1499px) {
+        @media (min-width: 1200px) and (max-width: 1499px) {
             font-size: .9em;
         }
     }
@@ -1194,7 +1204,6 @@ div[role="presentation"] {
     }
 }
 
-
 .nextra-code-block {
     margin: 0;
     width: 100%;
@@ -1206,5 +1215,67 @@ div[role="presentation"] {
     }
     pre {
         margin: 0;
+    }
+}
+
+#topButton {
+    display: block;
+    position: fixed;
+    bottom: 60px;
+    right: 30px;
+    z-index: 99;
+    font-size: 18px !important;
+    cursor: pointer;
+    border-radius: 7px !important;
+    margin: 0 !important;
+    outline: none !important;
+    border: none !important;
+    -webkit-backdrop-filter: blur(3px);
+    backdrop-filter: blur(3px);
+    padding: 0.3em 0.7em !important;
+    min-width: 35px !important;
+    min-height: 35px !important;
+    font-weight: 700 !important;
+    color: var(--bg) !important;
+    font-family: "montserrat" !important;
+    -webkit-animation: fadeIn 1000ms ease-in-out;
+    animation: fadeIn 1000ms ease-in-out;
+    background: var(--font) !important;
+    transition: var(--transition) !important;
+    .upArrow {
+        bottom: -3px;
+        position: relative;
+        transition: var(--transition) !important;
+    }
+    &:active {
+        background: black;
+        box-shadow: inset 0px 0px 3px rgba(0, 0, 0, 0.8);
+        color: var(--secondary);
+    }
+    &:hover {
+        background: var(--secondary) !important;
+        color: var(--bg) !important;
+        &::before {
+        transform: scale(1, 1);
+        opacity: 1;
+        }
+    }
+    &::before {
+        content: "";
+        position: absolute;
+        opacity: 0;
+        width: 100%;
+        height: 100%;
+        background: var(--primary);
+        border-radius: 7px !important;
+        top: 0;
+        left: 0;
+        z-index: -1;
+        transform-origin: top;
+        transform: scale(0, 1);
+        transition: var(--transition);
+    }
+    @media (max-width: 992px) {
+        bottom: 10px;
     }
 }
